to fazendo algo..........

+++ towerdefence/Assets/health.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class health : MonoBehaviour
 {
-    [SerializeField] private int hitponints = 2;
+    [SerializeField] private float hitponints = 2;
     private bool isDestroyed = false;
     [SerializeField] private int currencyWorth = 50;
 
-    public void takedamaga(int dmg)
+    public void takedamaga(float dmg)
     {
         hitponints-=dmg;
         if (hitponints <= 0&& !isDestroyed)
